moved logout button to settings
